hotfix: 파일 수정시 예외처리 추가 - 파일 경로가 tmp로 시작하지 않는 경우, 새로운 파일이 아니라고 판단하여 null 반환 - 파일 수정 시, 파일 삭제 로직 마지막으로 이동

+++ backend/src/main/java/com/ssafy/a208/domain/space/service/ArticleFileService.java
@@ -77,13 +77,17 @@ public class ArticleFileService {
             return null;
         }
 
+        if (!filePath.startsWith("tmp")) {
+            return null;
+        }
+
         Optional<ArticleFile> file = articleFileReader.getArticleFileById(article.getId());
         if (file.isPresent()) {
             ArticleFile existing = file.get();
             String destPath = extractFilePath(filePath);
-            s3Service.deleteFile(existing.getFilePath());
             FileMetaData metaData = s3Service.getFileMetadata(destPath);
             existing.updateFile(metaData);
+            s3Service.deleteFile(existing.getFilePath());
             return destPath;
         }
